Revert to simple chat component and move onboarding header into it
Reverts to SimpleChatInterface and incorporates the OnboardingChat header from onboarding.tsx page.
