+++ client/src/pages/onboarding.tsx
@@ -148,7 +148,6 @@ export default function Onboarding() {
 
   return (
     <div className="min-h-screen bg-gray-50 dark:bg-gray-900 flex flex-col overflow-hidden">
-      {!shouldShowDashboard && <Header />}
       <SimpleChatInterface 
         sessionId={sessionId} 
         userStatus={userStatus}
